# docs/openapi.yaml
...
openapi: 3.0.4
info:
  title: dployr Base API
  version: 1.0.0
  description: |
    Your app, your server, your rules!

    ## Authentication
    Authenticate via OAuth2.0 (Google, GitHub, Microsoft) or email OTP.
    Sessions are cookie-based.

    ## Authorization
    Role-based access control:
    - **Owner**: Full control, manages admins and billing
    - **Admin**: User management, infrastructure
    - **Developer**: Deploy apps, manage integrations
    - **Viewer**: Read-only access

    ## Rate Limiting
    - Global: 100 requests/minute per user or IP on `/v1/*`
    - Strict: 10 requests/minute on sensitive endpoints
    - Returns `429` with `Retry-After` and `X-RateLimit-*` headers on breach
servers:
  - url: https://base.dployr.dev
    description: Development
  - url: https://base.dployr.io
    description: Production
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              example: Resource not found
            code:
              type: string
              description: |
                Machine-readable error code. Known values:
                - `request.missing_params`
                - `request.bad_request`
                - `request.invalid_otp`
                - `request.unprocessable_entity`
                - `request.too_many_requests`
                - `request.internal_server_error`
                - `validation.missing_fields`
                - `validation.invalid_format`
                - `auth.bad_session`
                - `auth.bad_token`
                - `auth.invalid_oauth_provider`
                - `auth.bad_oauth_state`
                - `permission.forbidden`
                - `permission.owner_role_required`
                - `permission.admin_role_required`
                - `permission.developer_role_required`
                - `permission.viewer_role_required`
                - `resource.missing_resource`
                - `resource.conflict`
                - `runtime.bad_webhook_signature`
                - `runtime.internal_server_error`
                - `runtime.node_not_connected`
                - `bootstrap.bootstrap_setup_failure`
                - `bootstrap.bootstrap_run_failure`
              example: resource.missing_resource
            helpLink:
              type: string
              format: uri
              example: https://status.dployr.io
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          nullable: true
          type: object
    PaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
              minimum: 1
            pageSize:
              type: integer
              minimum: 1
            totalItems:
              type: integer
              minimum: 0
            totalPages:
              type: integer
              minimum: 0
            hasNextPage:
              type: boolean
            hasPreviousPage:
              type: boolean
...
paths:
...
  /v1/runtime/compatibility/check:
    post:
      operationId: checkCompatibility
      tags: [Runtime]
      summary: Check dployrd compatibility
      description: |
        Called by the daemon on startup to verify its `compatibility_date` meets
        the base server requirement. Public — no auth required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [compatibilityDate]
              properties:
                compatibilityDate:
                  type: string
                  pattern: '^\d{4}-\d{2}-\d{2}$'
                  example: "2025-12-28"
                version:
                  type: string
                  example: v0.4.8
      responses:
        '200':
          description: Compatibility result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          compatible:
                            type: boolean
                          compatibilityDate:
                            type: string
                          version:
                            type: string
                          latestVersion:
                            type: string
                          upgradeLevel:
                            type: string
                            enum: [major, minor, patch, none]
                          requiredCompatibilityDate:
                            type: string
                          message:
                            type: string
              examples:
                compatible:
                  value:
                    success: true
                    data:
                      compatible: true
                      compatibilityDate: "2025-12-28"
                      version: v0.4.8
                      latestVersion: v0.4.8
                      upgradeLevel: none
                      requiredCompatibilityDate: "2025-12-28"
                      message: dployrd is supported
                incompatible:
                  value:
                    success: true
                    data:
                      compatible: false
                      compatibilityDate: "2025-10-01"
                      version: v0.3.0
                      latestVersion: v0.4.8
                      upgradeLevel: major
                      requiredCompatibilityDate: "2025-12-28"
                      message: "dployrd requires compatibility_date 2025-12-28 (received: 2025-10-01)"
        '400':
          description: Invalid date format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runtime/versions:
    get:
      operationId: listRuntimeVersions
      tags: [Runtime]
      summary: List available dployrd versions from GitHub releases
      security: []
      parameters:
        - in: query
          name: showPreReleases
          schema:
            type: boolean
          description: Include prerelease versions
      responses:
        '200':
          description: Available versions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          latest:
                            type: string
                            nullable: true
                          oldestSupportedVersion:
                            type: string
                            nullable: true
                          versions:
                            type: array
                            items:
                              type: string
                          includePreReleases:
                            type: boolean
        '500':
          description: Failed to fetch from GitHub
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/runtime/events:
    get:
      operationId: getRuntimeEvents
      tags: [Runtime]
      summary: Get runtime events for a cluster
      description: Requires viewer role.
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: clusterId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
        - in: query
          name: type
          schema:
            type: string
          description: Filter by event type code
        - in: query
          name: search
          schema:
            type: string
          description: Free-text search across event payloads
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest]
          description: Sort order. Default is newest.
        - in: query
          name: window
          schema:
            type: string
            enum: [all, 24h, 7d, 30d]
          description: Time window filter. Default is all.
        - in: query
          name: actor
          schema:
            type: string
          description: Filter by actor (user ID or 'headless')
      responses:
        '200':
          description: Paginated events
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PaginatedResponse'
        '400':
          description: Missing clusterId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Nodes ─────────────────────────────────────────────────────────────────
...